requirements.txt
# Function dependencies, for example:
# package>=version

line-bot-sdk
requests
git+https://github.com/ozgur/python-firebase
google.generativeai
pillow
grpcio